pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-ai-ml>=1.30.0",
    "fastapi>=0.121.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]
